@@ -5,9 +5,10 @@
 // Scalars can be used interchangeably between groups.
 //
 // These groups have the same order equal to:
-//  Order = 0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001
 //
-// Serialization Format
+//	Order = 0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001
+//
+// # Serialization Format
 //
 // Elements of G1 and G2 can be encoded in uncompressed form (the x-coordinate
 // followed by the y-coordinate) or in compressed form (just the x-coordinate).
@@ -31,24 +32,23 @@
 // is the lexicographically largest of the two associated with the encoded
 // x-coordinate.
 //
-//  |----------------------------------------------------|
-//  |                Serialization Format                |
-//  |-----|-------|-------|---------------|--------------|
-//  | MSB | MSB-1 | MSB-2 |  Description  | Encoding     |
-//  |-----|-------|-------|---------------|--------------|
-//  |  0  |   X   |   X   | Uncompressed  |  e || x || y |
-//  |  1  |   X   |   X   | Compressed    |  e || x      |
-//  |-----|-------|-------|---------------|--------------|
-//  |  X  |   0   |   X   | Non-Infinity  |  e || x || y |
-//  |  X  |   1   |   X   | Infinity      |  e || 0 || 0 |
-//  |-----|-------|-------|---------------|--------------|
-//  |     |       |       | Compressed,   |              |
-//  |  1  |   0   |   1   | Non-Infinity, |  e || x      |
-//  |     |       |       | Big y-coord   |              |
-//  |-----|-------|-------|---------------|--------------|
-//  |     |       |       | Compressed,   |              |
-//  |  1  |   0   |   0   | Non-Infinity, |  e || x      |
-//  |     |       |       | Small y-coord |              |
-//  |----------------------------------------------------|
-//
+//	|----------------------------------------------------|
+//	|                Serialization Format                |
+//	|-----|-------|-------|---------------|--------------|
+//	| MSB | MSB-1 | MSB-2 |  Description  | Encoding     |
+//	|-----|-------|-------|---------------|--------------|
+//	|  0  |   X   |   X   | Uncompressed  |  e || x || y |
+//	|  1  |   X   |   X   | Compressed    |  e || x      |
+//	|-----|-------|-------|---------------|--------------|
+//	|  X  |   0   |   X   | Non-Infinity  |  e || x || y |
+//	|  X  |   1   |   X   | Infinity      |  e || 0 || 0 |
+//	|-----|-------|-------|---------------|--------------|
+//	|     |       |       | Compressed,   |              |
+//	|  1  |   0   |   1   | Non-Infinity, |  e || x      |
+//	|     |       |       | Big y-coord   |              |
+//	|-----|-------|-------|---------------|--------------|
+//	|     |       |       | Compressed,   |              |
+//	|  1  |   0   |   0   | Non-Infinity, |  e || x      |
+//	|     |       |       | Small y-coord |              |
+//	|----------------------------------------------------|
 package bls12381
