@@ -64,7 +64,8 @@ func (z *Fp) CMov(x, y *Fp, b int) {
 }
 
 // FpOrder is the order of the base field for towering returned as a big-endian slice.
-//  FpOrder = 0x1a0111ea397fe69a4b1ba7b6434bacd764774b84f38512bf6730d2a0f6b0f6241eabfffeb153ffffb9feffffffffaaab.
+//
+//	FpOrder = 0x1a0111ea397fe69a4b1ba7b6434bacd764774b84f38512bf6730d2a0f6b0f6241eabfffeb153ffffb9feffffffffaaab.
 func FpOrder() []byte { o := fpOrder; return o[:] }
 
 // ExpVarTime calculates z=x^n, where n is the exponent in big-endian order.
