@@ -87,20 +87,20 @@ func unmarshalContext(raw []byte) (*encdecContext, error) {
 //
 // enum { sealer(0), opener(1) } HpkeRole;
 //
-// struct {
-//     HpkeKemId kem_id;   // draft-irtf-cfrg-hpke-07
-//     HpkeKdfId kdf_id;   // draft-irtf-cfrg-hpke-07
-//     HpkeAeadId aead_id; // draft-irtf-cfrg-hpke-07
-//     opaque exporter_secret<0..255>;
-//     opaque key<0..255>;
-//     opaque base_nonce<0..255>;
-//     opaque seq<0..255>;
-// } HpkeContext;
+//	struct {
+//	    HpkeKemId kem_id;   // draft-irtf-cfrg-hpke-07
+//	    HpkeKdfId kdf_id;   // draft-irtf-cfrg-hpke-07
+//	    HpkeAeadId aead_id; // draft-irtf-cfrg-hpke-07
+//	    opaque exporter_secret<0..255>;
+//	    opaque key<0..255>;
+//	    opaque base_nonce<0..255>;
+//	    opaque seq<0..255>;
+//	} HpkeContext;
 //
-// struct {
-//   HpkeRole role = 0; // sealer
-//   HpkeContext context;
-// } HpkeSealer;
+//	struct {
+//	  HpkeRole role = 0; // sealer
+//	  HpkeContext context;
+//	} HpkeSealer;
 func (c *sealContext) MarshalBinary() ([]byte, error) {
 	rawContext, err := c.encdecContext.marshal()
 	if err != nil {
@@ -125,10 +125,10 @@ func UnmarshalSealer(raw []byte) (Sealer, error) {
 // below. (Expressed in TLS syntax.) Note that this format is not defined by the
 // HPKE standard.
 //
-// struct {
-//   HpkeRole role = 1; // opener
-//   HpkeContext context;
-// } HpkeOpener;
+//	struct {
+//	  HpkeRole role = 1; // opener
+//	  HpkeContext context;
+//	} HpkeOpener;
 func (c *openContext) MarshalBinary() ([]byte, error) {
 	rawContext, err := c.encdecContext.marshal()
 	if err != nil {
