@@ -5,6 +5,7 @@
 // implementation [2].
 //
 // References:
-//  [1] https://frodokem.org/files/FrodoKEM-specification-20210604.pdf
-//  [2] https://github.com/PQClean/PQClean/tree/master/crypto_kem/frodokem640shake/clean
+//
+//	[1] https://frodokem.org/files/FrodoKEM-specification-20210604.pdf
+//	[2] https://github.com/PQClean/PQClean/tree/master/crypto_kem/frodokem640shake/clean
 package frodo
