@@ -1,15 +1,19 @@
 // Package schemes contains a register of KEM schemes.
 //
-// Schemes Implemented
+// # Schemes Implemented
 //
 // Based on standard elliptic curves:
-//  HPKE_KEM_P256_HKDF_SHA256, HPKE_KEM_P384_HKDF_SHA384, HPKE_KEM_P521_HKDF_SHA512
+//
+//	HPKE_KEM_P256_HKDF_SHA256, HPKE_KEM_P384_HKDF_SHA384, HPKE_KEM_P521_HKDF_SHA512
+//
 // Based on standard Diffie-Hellman functions:
-//  HPKE_KEM_X25519_HKDF_SHA256, HPKE_KEM_X448_HKDF_SHA512
+//
+//	HPKE_KEM_X25519_HKDF_SHA256, HPKE_KEM_X448_HKDF_SHA512
+//
 // Post-quantum kems:
-//  FrodoKEM-640-SHAKE
-//  Kyber512, Kyber768, Kyber1024
-//  SIKEp434, SIKEp503, SIKEp751
+//
+//	FrodoKEM-640-SHAKE
+//	Kyber512, Kyber768, Kyber1024
 package schemes
 
 import (
@@ -22,9 +26,6 @@ import (
 	"github.com/cloudflare/circl/kem/kyber/kyber1024"
 	"github.com/cloudflare/circl/kem/kyber/kyber512"
 	"github.com/cloudflare/circl/kem/kyber/kyber768"
-	"github.com/cloudflare/circl/kem/sike/sikep434"
-	"github.com/cloudflare/circl/kem/sike/sikep503"
-	"github.com/cloudflare/circl/kem/sike/sikep751"
 )
 
 var allSchemes = [...]kem.Scheme{
@@ -37,9 +38,6 @@ var allSchemes = [...]kem.Scheme{
 	kyber512.Scheme(),
 	kyber768.Scheme(),
 	kyber1024.Scheme(),
-	sikep434.Scheme(),
-	sikep503.Scheme(),
-	sikep751.Scheme(),
 	hybrid.Kyber512X25519(),
 	hybrid.Kyber768X25519(),
 	hybrid.Kyber768X448(),
