@@ -9,7 +9,7 @@
 // subpackage.  For instance, Dilithium2 (the recommended mode)
 // can be found in
 //
-//  github.com/cloudflare/circl/sign/dilithium/mode2
+//	github.com/cloudflare/circl/sign/dilithium/mode2
 //
 // If your choice for mode is fixed compile-time, use the subpackages.
 // This package provides a convenient wrapper around all of the subpackages
@@ -18,8 +18,8 @@
 // The authors of Dilithium recommend to combine it with a "pre-quantum"
 // signature scheme.  The packages
 //
-//  github.com/cloudflare/circl/sign/eddilithium2
-//  github.com/cloudflare/circl/sign/eddilithium3
+//	github.com/cloudflare/circl/sign/eddilithium2
+//	github.com/cloudflare/circl/sign/eddilithium3
 //
 // implement such hybrids of Dilithium2 with Ed25519 respectively and
 // Dilithium3 with Ed448.  These packages are a drop in replacements for the
